Sync linting configs across root, BushelCloud, CelestraCloud, MistDemo and fix violations
- Sync lint.sh, header.sh, .swift-format, .swiftlint.yml, .periphery.yml across all 4 directories
- Add no_unchecked_sendable custom rule, one_declaration_per_file, type_name config to all .swiftlint.yml
- Create full linting setup for MistDemo (Scripts, configs, mise.toml)
- Update header.sh to heredoc template with % escape security fix
- Disable FileScopedDeclarationPrivacy in .swift-format (conflicts with swiftlint)
- Fix all swift-format and swiftlint violations across all 4 projects
