+++ Sources/CelestraCloud/Services/FeedUpdateResult.swift
@@ -3,11 +3,11 @@
 //  CelestraCloud
 //
 //  Created by Leo Dion.
-//  Copyright © 2025 BrightDigit.
+//  Copyright © 2026 BrightDigit.
 //
 //  Permission is hereby granted, free of charge, to any person
 //  obtaining a copy of this software and associated documentation
-//  files (the “Software”), to deal in the Software without
+//  files (the "Software"), to deal in the Software without
 //  restriction, including without limitation the rights to use,
 //  copy, modify, merge, publish, distribute, sublicense, and/or
 //  sell copies of the Software, and to permit persons to whom the
@@ -17,7 +17,7 @@
 //  The above copyright notice and this permission notice shall be
 //  included in all copies or substantial portions of the Software.
 //
-//  THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 //  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 //  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 //  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -29,10 +29,23 @@
 
 /// Result of processing a single feed update
 internal enum FeedUpdateResult: Sendable, Equatable {
+  // MARK: - Cases
+
   case success(articlesCreated: Int, articlesUpdated: Int)
   case notModified
   case skipped(reason: String)
   case error(message: String)
+
+  // MARK: - Subtypes
+
+  internal enum SimpleStatus {
+    case success
+    case notModified
+    case skipped
+    case error
+  }
+
+  // MARK: - Properties
 
   /// Simple status for backward compatibility
   internal var simpleStatus: SimpleStatus {
@@ -46,12 +59,5 @@ internal enum FeedUpdateResult: Sendable, Equatable {
     case .error:
       return .error
     }
-  }
-
-  internal enum SimpleStatus {
-    case success
-    case notModified
-    case skipped
-    case error
   }
 }
