+++ Tests/CelestraCloudTests/Errors/CelestraErrorTests.swift
@@ -44,8 +44,8 @@ internal struct CelestraErrorTests {
   }
 
   @Test("RSS fetch failed is retriable")
-  internal func testRSSFetchFailedRetriable() {
-    let url = URL(string: "https://example.com/feed.xml")!
+  internal func testRSSFetchFailedRetriable() throws {
+    let url = try #require(URL(string: "https://example.com/feed.xml"))
     let underlyingError = NSError(domain: "Test", code: 1)
     let error = CelestraError.rssFetchFailed(url, underlying: underlyingError)
 
@@ -131,132 +131,5 @@ internal struct CelestraErrorTests {
     let error = CelestraError.cloudKitError(ckError)
 
     #expect(error.isRetriable == true)
-  }
-
-  // MARK: - Error Description Tests
-
-  @Test("Quota exceeded has description")
-  internal func testQuotaExceededDescription() {
-    let error = CelestraError.quotaExceeded
-
-    #expect(error.errorDescription?.contains("quota") == true)
-    #expect(error.errorDescription?.contains("exceeded") == true)
-  }
-
-  @Test("Network unavailable has description")
-  internal func testNetworkUnavailableDescription() {
-    let error = CelestraError.networkUnavailable
-
-    #expect(error.errorDescription?.contains("Network") == true)
-    #expect(error.errorDescription?.contains("unavailable") == true)
-  }
-
-  @Test("Permission denied has description")
-  internal func testPermissionDeniedDescription() {
-    let error = CelestraError.permissionDenied
-
-    #expect(error.errorDescription?.contains("Permission") == true)
-    #expect(error.errorDescription?.contains("denied") == true)
-  }
-
-  @Test("Invalid feed data includes reason")
-  internal func testInvalidFeedDataDescription() {
-    let error = CelestraError.invalidFeedData("Malformed XML")
-
-    #expect(error.errorDescription?.contains("Invalid feed data") == true)
-    #expect(error.errorDescription?.contains("Malformed XML") == true)
-  }
-
-  @Test("Record not found includes record name")
-  internal func testRecordNotFoundDescription() {
-    let error = CelestraError.recordNotFound("feed-abc123")
-
-    #expect(error.errorDescription?.contains("Record not found") == true)
-    #expect(error.errorDescription?.contains("feed-abc123") == true)
-  }
-
-  @Test("RSS fetch failed includes URL")
-  internal func testRSSFetchFailedDescription() {
-    let url = URL(string: "https://example.com/feed.xml")!
-    let underlyingError = NSError(domain: "Test", code: 1, userInfo: [
-      NSLocalizedDescriptionKey: "Connection timeout",
-    ])
-    let error = CelestraError.rssFetchFailed(url, underlying: underlyingError)
-
-    #expect(error.errorDescription?.contains("example.com/feed.xml") == true)
-    #expect(error.errorDescription?.contains("Failed to fetch") == true)
-  }
-
-  @Test("Batch operation failed includes error count")
-  internal func testBatchOperationFailedDescription() {
-    let errors: [any Error] = [
-      NSError(domain: "Test", code: 1),
-      NSError(domain: "Test", code: 2),
-      NSError(domain: "Test", code: 3),
-    ]
-    let error = CelestraError.batchOperationFailed(errors)
-
-    #expect(error.errorDescription?.contains("3") == true)
-    #expect(error.errorDescription?.contains("Batch operation failed") == true)
-  }
-
-  // MARK: - Recovery Suggestion Tests
-
-  @Test("Quota exceeded has recovery suggestion")
-  internal func testQuotaExceededRecoverySuggestion() {
-    let error = CelestraError.quotaExceeded
-
-    #expect(error.recoverySuggestion != nil)
-    #expect(error.recoverySuggestion?.contains("Wait") == true)
-    #expect(error.recoverySuggestion?.contains("quota") == true)
-  }
-
-  @Test("Network unavailable has recovery suggestion")
-  internal func testNetworkUnavailableRecoverySuggestion() {
-    let error = CelestraError.networkUnavailable
-
-    #expect(error.recoverySuggestion != nil)
-    #expect(error.recoverySuggestion?.contains("connection") == true)
-  }
-
-  @Test("RSS fetch failed has recovery suggestion")
-  internal func testRSSFetchFailedRecoverySuggestion() {
-    let url = URL(string: "https://example.com/feed.xml")!
-    let underlyingError = NSError(domain: "Test", code: 1)
-    let error = CelestraError.rssFetchFailed(url, underlying: underlyingError)
-
-    #expect(error.recoverySuggestion != nil)
-    #expect(error.recoverySuggestion?.contains("feed URL") == true)
-  }
-
-  @Test("Permission denied has recovery suggestion")
-  internal func testPermissionDeniedRecoverySuggestion() {
-    let error = CelestraError.permissionDenied
-
-    #expect(error.recoverySuggestion != nil)
-    #expect(error.recoverySuggestion?.contains("permissions") == true)
-  }
-
-  @Test("Invalid feed data has recovery suggestion")
-  internal func testInvalidFeedDataRecoverySuggestion() {
-    let error = CelestraError.invalidFeedData("Invalid XML")
-
-    #expect(error.recoverySuggestion != nil)
-    #expect(error.recoverySuggestion?.contains("RSS") == true)
-  }
-
-  @Test("Record not found has no recovery suggestion")
-  internal func testRecordNotFoundNoRecoverySuggestion() {
-    let error = CelestraError.recordNotFound("feed-123")
-
-    #expect(error.recoverySuggestion == nil)
-  }
-
-  @Test("CloudKit error has no recovery suggestion")
-  internal func testCloudKitErrorNoRecoverySuggestion() {
-    let ckError = CloudKitError.httpError(statusCode: 500)
-    let error = CelestraError.cloudKitError(ckError)
-
-    #expect(error.recoverySuggestion == nil)
   }
 }
